Add local frontend and backend test support
